selectionError string :
For when there's an error with the selected option. Probably from loading the filter, probably because a mod was turned off. I imagined modded defs should be there, but right now it's just the ListModded filter, with the modcontentpack itself that sets selectionError. Also should apply to named zones.

Renaming unspecific Callback methods to PostChosen, so they don't need to assign Sel. These two changes were more intertwined but I rolled back and redesigned a lot.