Cleaner owner of filters with the IFilterOwner/IFilterOwnerAdder interface. Both the root FindDescription and ListFilterGroup/Selection are Owners where all but Selection are also IFilterOwnerAdders.
So instead of each ListFilter holding onto the root FindDescription, they hold onto possibly their owner ListFilterGroup, which then can access the FindDescription. So they can tell if they are a top-level filter or not if their owner is a FindDescription.

Also cleaner code to add new filters to the list