Can't drop group inside its own filters.
Also thinking that dragging filters to add them bypasses setting the parent. So. Let's overhaul code so FilterOwner.Add() is the only way to add a filter, and that sets the parent. That saves a lot of passing IFilterHolder to Clone, because the cloned filer is just going to be added immediately after.